+++ actix-web-lab/src/middleware.rs
@@ -8,6 +8,7 @@ pub use crate::{
     load_shed::LoadShed,
     middleware_from_fn::{from_fn, MiddlewareFn, Next},
     middleware_map_response::{map_response, MapResMiddleware},
+    middleware_map_response_body::{map_response_body, MapResBodyMiddleware},
     normalize_path::NormalizePath,
     panic_reporter::PanicReporter,
     redirect_to_https::RedirectHttps,