+++ actix-web-lab/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased - 2022-xx-xx
 
 - Add `middleware::map_response()` for mapping responses with an async function.
+- Add `middleware::map_response_body()` for mapping response bodies with an async function.
 
 ## 0.18.8 - 2022-12-10